Installed navigation dependencies
