+++ src/Navigator/AppNavigator.tsx
@@ -1,0 +1,26 @@
+import * as React from "react";
+import { View, StyleSheet, Text } from "react-native";
+import { StatusBar } from "expo-status-bar";
+import { NavigationContainer } from "@react-navigation/native";
+
+function AppNavigator() {
+  return (
+    <NavigationContainer>
+      <View style={styles.container}>
+        <Text>Open up App.tsx to start working on your app!</Text>
+        <StatusBar style="auto" />
+      </View>
+    </NavigationContainer>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
+
+export default AppNavigator;
